Feature/terraform deploy (#26)
* add in vpc_id and security group lookups

* removed pyc files

* /version 0.9.0-alpha.4

* update terraform deploy

* update to have default message config passed in as a configuration

* update forge py

* moved regression and memory profiler scripts

* fix pylint, add changelog, and update forge-py

* update default message config with example values

---------
