+++ .github/workflows/build.yml
@@ -74,6 +74,7 @@ jobs:
           echo "software_version=${new_version}" >> $GITHUB_ENV
           echo "new_version=${new_version}" >> $GITHUB_ENV
           echo "Github REF: ${{ github.ref }}"
+          echo "TARGET_ENV_UPPERCASE=SIT" >> $GITHUB_ENV
 
       # Alpha Logic - Use the project version number and add -alpha.1 or bump
       # alpha number
@@ -173,11 +174,16 @@ jobs:
           echo "VENUE: ${{ env.venue }}"
           echo "Target Env Uppercase: ${{ env.TARGET_ENV_UPPERCASE }}"
 
+          original_env_value="${TARGET_ENV_UPPERCASE}"
+          lowercase_value=$(echo "${original_env_value}" | tr '[:upper:]' '[:lower:]')
+          echo "TARGET_ENV_LOWERCASE=${lowercase_value}" >> $GITHUB_ENV
+
       # NOTE: This step is platform-specific
       # Update the version number in the application package itself
       - name: Update version number in the application package
         run: |
           poetry version ${{ env.the_version }}
+
 
       #########################################################################
       # Install & Test & Snyk
@@ -291,3 +297,51 @@ jobs:
         run: |
           pip install tenacity logging 
           python3 ${GITHUB_WORKSPACE}/.github/workflows/wait-for-pypi.py ${{env.pyproject_name}}==${{ env.software_version }}
+
+
+      #########################################################################
+      # Deploy to AWS via Terraform
+      #########################################################################
+      - name: Deploy Env Override
+        if: |
+          github.event.head_commit.message == '/deploy sit' ||
+          github.event.head_commit.message == '/deploy uat'
+        run: |
+          message="${{ github.event.head_commit.message }}"
+          trimmed_message=${message:1}  # Remove leading slash
+          override_env=$(echo "$trimmed_message" | grep -oE '[^[:space:]]+$')
+          override_env_upper=$(echo "$trimmed_message" | awk '{print toupper($NF)}')
+          echo "THE_ENV=${override_env}" >> $GITHUB_ENV
+          echo "TARGET_ENV_LOWERCASE=${override_env}" >> $GITHUB_ENV
+          echo "TARGET_ENV_UPPERCASE=${override_env_upper}" >> $GITHUB_ENV
+
+      - uses: hashicorp/setup-terraform@v3
+        with:
+          terraform_version: 1.5.3
+
+      - name: Deploy Terraform
+        if: |
+          github.ref == 'refs/heads/develop' ||
+          github.ref == 'refs/heads/main'    ||
+          startsWith(github.ref, 'refs/heads/release') ||
+          github.event.head_commit.message == '/deploy sit' ||
+          github.event.head_commit.message == '/deploy uat'
+        working-directory: terraform-deploy/
+        env:
+          AWS_ACCESS_KEY_ID:       ${{ secrets[format('AWS_ACCESS_KEY_ID_SERVICES_{0}', env.TARGET_ENV_UPPERCASE)] }}
+          AWS_SECRET_ACCESS_KEY:   ${{ secrets[format('AWS_SECRET_ACCESS_KEY_SERVICES_{0}', env.TARGET_ENV_UPPERCASE)] }}
+          AWS_ACCOUNT_ID:          ${{ secrets[format('AWS_ACCOUNT_ID_SERVICES_{0}', env.TARGET_ENV_UPPERCASE)] }}
+          AWS_DEFAULT_REGION:      us-west-2
+
+          TF_VAR_permissions_boundary_arn: ${{ secrets[format('PERMISSIONS_BOUNDARY_ARN_{0}', env.TARGET_ENV_UPPERCASE)] }}
+          TF_VAR_buckets_name: ${{ secrets[format('BUCKET_{0}', env.TARGET_ENV_UPPERCASE)] }}
+          TF_VAR_system_bucket: ${{ secrets[format('SYSTEM_BUCKET_{0}', env.TARGET_ENV_UPPERCASE)] }}
+          TF_VAR_dmrpp_url: ${{ secrets.DMRPP_URL }}
+          TF_VAR_aws_security_group_ids: ${{ secrets[format('SECURITY_GROUP_IDS_{0}', env.TARGET_ENV_UPPERCASE)] }}
+
+        run: |
+          curl -L -o metadata-aggregator.zip https://github.com/podaac/cumulus-metadata-aggregator/releases/download/v8.6.0/cumulus-metadata-aggregator-8.6.0.zip
+          python3 override.py "${{ env.TARGET_ENV_LOWERCASE }}"
+          source bin/config.sh ${{ env.TARGET_ENV_LOWERCASE }}
+          terraform plan -var-file=tfvars/"${{ env.TARGET_ENV_LOWERCASE }}".tfvars -var="app_version=${{ env.the_version }}" -out="tfplan"
+          terraform apply -auto-approve tfplan
