+++ podaac/hitide_backfill_tool/default_message_config.json
@@ -1,10 +1,27 @@
 {
-   "uat":{
-      "buckets":{
-         "internal":{
-            "name":"test-internal",
-            "type":"internal"
-         }
-      }
-   }
+  "sit": {
+    "buckets": {
+      "internal": {
+        "name": "internal_bucket",
+        "type": "internal"
+      },
+    },
+    "stack": "sit",
+    "cmr": {
+      "clientId": "clientId",
+      "cmrEnvironment": "cmrEnvironment",
+      "cmrLimit": 100,
+      "cmrPageSize": 50,
+      "oauthProvider": "oauthProvider",
+      "passwordSecretName": "passwordSecretName",
+      "provider": "provider",
+      "username": "username"
+    },
+    "distribution_endpoint": "https://example.com/",
+    "launchpad": {
+      "api": "https://launchpadapi.com",
+      "certificate": "certificate.pfx",
+      "passphraseSecretName": "passphraseSecretName"
+    }
+  }
 }
