+++ terraform-deploy/ecs_cluster_instance_autoscaling_cf_template.yml.tmpl
@@ -13,8 +13,8 @@ Resources:
         - DeviceName: "/dev/xvdcz"
           Ebs:
             DeleteOnTermination: true
+            Encrypted: true
             VolumeSize: ${docker_volume_size}
-            VolumeType: gp3
 %{ if key_name != null ~}
       KeyName: ${key_name}
 %{ endif ~}
@@ -26,20 +26,19 @@ Resources:
           --==BOUNDARY==
           Content-Type: text/cloud-boothook; charset="us-ascii"
 
-%{ if docker_storage_driver == "devicemapper" ~}
           vgcreate docker /dev/xvdcz
 
           lvcreate -n docker-data -L${docker_volume_create_size}G docker
 
+          mkfs.xfs /dev/docker/docker-data
+          mkdir /docker-data
+          mount /dev/docker/docker-data /docker-data
+
           sed -i '/^\s*DOCKER_STORAGE_OPTIONS=/d' /etc/sysconfig/docker-storage
-          echo 'DOCKER_STORAGE_OPTIONS="--storage-driver devicemapper --storage-opt dm.datadev=/dev/docker/docker-data"' >> /etc/sysconfig/docker-storage
+          echo 'DOCKER_STORAGE_OPTIONS="--storage-driver overlay2' >> /etc/sysconfig/docker-storage
 
           sed -i '/^\s*OPTIONS=/d' /etc/sysconfig/docker
-          echo 'OPTIONS="-D --default-ulimit nofile=1024:4096 --storage-opt dm.basesize=${docker_volume_size}G --storage-opt dm.datadev=/dev/docker/docker-data"' >> /etc/sysconfig/docker
-%{ else ~}
-          sed -i '/^\s*DOCKER_STORAGE_OPTIONS=/d' /etc/sysconfig/docker-storage
-          echo 'DOCKER_STORAGE_OPTIONS="--storage-driver ${docker_storage_driver}"' >> /etc/sysconfig/docker-storage
-%{ endif ~}
+          echo 'OPTIONS="--default-ulimit nofile=1024:4096 --data-root=/docker-data"' >> /etc/sysconfig/docker
 
 %{ if include_docker_cleanup_cronjob == true ~}
           echo '* * * * * sudo sh -c "docker ps -q | xargs docker inspect --format='\{{.State.Pid}}' | xargs -IZ fstrim /proc/Z/root/"' | crontab -
