+++ terraform-deploy/environments/sit.env
@@ -1,0 +1,2 @@
+export REGION=us-west-2
+export BUCKET=podaac-services-sit-terraform
