+++ terraform-deploy/forge.tf
@@ -9,8 +9,8 @@ module "forge_module" {
     footprint_output_dir    = "dataset-metadata"
     lambda_role = aws_iam_role.iam_execution.arn
     layers = [aws_lambda_layer_version.cumulus_message_adapter.arn]
-    security_group_ids = var.security_group_ids
-    subnet_ids = var.subnet_ids
+    security_group_ids = [var.aws_security_group_ids]
+    subnet_ids = data.aws_subnets.private.ids
     memory_size = 512
     timeout = 900
     profile = var.profile
