+++ terraform-deploy/hyrax-metadata-updates.tf
@@ -21,8 +21,8 @@ resource "aws_lambda_function" "hyrax_metadata_updates_task" {
   }
 
   vpc_config {
-    subnet_ids = var.subnet_ids
-    security_group_ids = var.security_group_ids
+    subnet_ids = data.aws_subnets.private.ids
+    security_group_ids = [var.aws_security_group_ids]
   }
 
   tags = {
