+++ terraform-deploy/postworkflow_normalizer.tf
@@ -5,8 +5,8 @@ module "postworkflow_normalizer_module" {
     region = var.region
     aws_profile = var.profile
     lambda_role = aws_iam_role.iam_execution.arn
-    security_group_ids = var.security_group_ids
-    subnet_ids = var.subnet_ids
+    security_group_ids = [var.aws_security_group_ids]
+    subnet_ids = data.aws_subnets.private.ids
     memory_size = 128
     timeout = 180
