+++ terraform-deploy/terraform_env/ops.json
@@ -3,7 +3,7 @@
    "tig_image":"ghcr.io/podaac/tig:0.12.0",
    "forge_source":"https://github.com/podaac/forge/releases/download/0.11.0/forge-0.11.0.zip",
    "forge_image":"ghcr.io/podaac/forge:0.11.0",
-   "forge_py_source" : "https://github.com/podaac/forge-py/releases/download/0.1.0/forge-py-terraform-0.1.0.zip",
-   "forge_py_image": "ghcr.io/podaac/forge-py:0.1.0",
+   "forge_py_source" : "https://github.com/podaac/forge-py/releases/download/0.2.0/forge-py-terraform-0.2.0.zip",
+   "forge_py_image": "ghcr.io/podaac/forge-py:0.2.0",
    "backfill_lambdas_source": "https://github.com/podaac/hitide-backfill-lambdas/releases/download/0.4.0/hitide-backfill-lambdas-0.4.0.zip"
 }
